rename slack_integrator to slack_poster
