+++ slack_poster/.gitignore
