+++ slack_poster/Gemfile
