+++ slack_poster/Rakefile
