+++ slack_poster/lib/slack_poster/network_request.rb
