+++ slack_poster/test/test_helper.rb
